+++ src/systems.rs
@@ -1,6 +1,6 @@
 //! The systems that power each [`InputManagerPlugin`](crate::plugin::InputManagerPlugin).
 
-use crate::pool::Pool;
+use crate::pool::RegeneratingPool;
 use crate::{charges::ChargeState, cooldown::CooldownState, Abilitylike};
 
 use bevy::ecs::prelude::*;
@@ -37,7 +37,7 @@ pub fn tick_cooldowns<A: Abilitylike>(
 }
 
 /// Regenerates the resource of the [`Pool`] type `P` based on the elapsed [`Time`].
-pub fn regenerate_resource_pool<P: Pool + Component + Resource>(
+pub fn regenerate_resource_pool<P: RegeneratingPool + Component + Resource>(
     mut query: Query<&mut P>,
     pool_res: Option<ResMut<P>>,
     time: Res<Time>,