+++ RELEASES.md
@@ -6,6 +6,8 @@
 
 - removed the required `new` method from the `Pool` trait: this method was overly restrictive, and prevented the construction of more complex pools with custom initialization parameters
   - `LifePool::new` and `ManaPool::new` methods have been added to the premade pools: do similarly for your own `Pool` types
+- the `Pool::ZERO` associated constant has been renamed to the clearer `Pool::MIN`.
+  - the `MaxPoolLessThanZero` error type has been renamed to `MaxPoolLessThanMin` to match.
 
 ## Version 0.5