+++ RELEASES.md
@@ -15,6 +15,7 @@
   - the `MaxPoolLessThanZero` error type has been renamed to `MaxPoolLessThanMin` to match.
 - the `Pool` trait has been split in two, with the regeneration-specific mechanics handled in `RegeneratingPool`, to make the construction of non-regenerating pools much more intuitive
 - added the `Pool::is_empty` and `Pool::is_full` helper methods to the `Pool` trait
+- added `Add`, `Sub`, `AddAssign` and `SubAssign` implementations to the premade `Life` and `Mana` types and their corresponding pools
 
 ## Version 0.5